Move majority of project files into src directory, and adjust imports accordingly, with new absolute src path mapping
