+++ src/constructs/saved-channels-api-routes.ts
@@ -6,11 +6,11 @@ import {
 } from "aws-cdk-lib/aws-lambda-nodejs";
 import { Construct } from "constructs";
 import { join } from "path";
-import * as lambda from "aws-cdk-lib/aws-lambda";
 import { SubnetType } from "aws-cdk-lib/aws-ec2";
 import { databaseConfig } from "../config/database";
 import { Duration } from "aws-cdk-lib";
 import { ApiRoutesProps } from "./http-api";
+import { Runtime } from "aws-cdk-lib/aws-lambda";
 
 export class SavedChannelApiRoutes extends Construct {
   constructor(scope: Construct, id: string, props: ApiRoutesProps) {
@@ -19,7 +19,7 @@ export class SavedChannelApiRoutes extends Construct {
     const { httpApi, authorizer, vpc } = props;
 
     const nodeJsFunctionProps: NodejsFunctionProps = {
-      runtime: lambda.Runtime.NODEJS_16_X,
+      runtime: Runtime.NODEJS_16_X,
       vpc: vpc,
       vpcSubnets: { subnetType: SubnetType.PRIVATE_ISOLATED },
       environment: databaseConfig,
