+++ src/lambdas/addSavedChannel.ts
@@ -1,8 +1,8 @@
 import { APIGatewayProxyEventV2, Handler } from "aws-lambda";
 import { is } from "superstruct";
-import { ChannelStruct } from "../types";
-import { insertChannel } from "../helpers/databaseQueries";
-import { database } from "../config/database";
+import { SavedChannelStruct } from "@/types";
+import { insertSavedChannel } from "@/helpers";
+import { database } from "@/config";
 
 export const handler: Handler = async function (event: APIGatewayProxyEventV2) {
   if (!event.body) {
@@ -10,26 +10,30 @@ export const handler: Handler = async function (event: APIGatewayProxyEventV2) {
       statusCode: 400,
       headers: { "Content-Type": "application/json" },
       body: {
-        message: "Bad request. Missing request body.",
+        message: "Bad request. Missing channel and/or user information.",
       },
     });
   }
 
-  const channelInformation = JSON.parse(event.body);
+  const savedChannelInformation = JSON.parse(event.body);
 
-  if (!is(channelInformation, ChannelStruct)) {
+  if (!is(savedChannelInformation, SavedChannelStruct)) {
     return JSON.stringify({
       statusCode: 400,
       headers: { "Content-Type": "application/json" },
       body: {
-        message: "Bad request. Invalid channel information.",
+        message: "Bad request. Invalid channel and/or user information.",
       },
     });
   }
 
   await database.connect();
 
-  await insertChannel(database, channelInformation);
+  await insertSavedChannel(
+    database,
+    savedChannelInformation.user_id,
+    savedChannelInformation.user_id
+  );
 
   await database.end();
 
@@ -37,7 +41,7 @@ export const handler: Handler = async function (event: APIGatewayProxyEventV2) {
     statusCode: 200,
     headers: { "Content-Type": "application/json" },
     body: {
-      message: "Channel added successfully",
+      message: "Saved channel added successfully",
     },
   });
 };
