+++ src/test/addChannel.test.ts
@@ -1,40 +1,41 @@
 import { APIGatewayProxyEventV2 } from "aws-lambda";
-import { handler } from "../lambdas/deleteChannel";
+import { handler } from "@/lambdas/addChannel";
 import { mockCallback, mockContext, mockEvent } from "./constants";
 
-const eventNoParams: APIGatewayProxyEventV2 = {
+const eventNoBody: APIGatewayProxyEventV2 = {
   ...mockEvent,
-  pathParameters: undefined,
+  body: undefined,
 };
 
-const eventBadChannelId: APIGatewayProxyEventV2 = {
+const eventBadChannel: APIGatewayProxyEventV2 = {
   ...mockEvent,
-  pathParameters: {
-    user_id: "1234",
-  },
+  body: JSON.stringify({
+    channelId: 123,
+    platform: "twitch",
+  }),
 };
 
-it("returns bad request for missing path params", async () => {
-  const response = await handler(eventNoParams, mockContext, mockCallback);
+it("returns bad request for missing body", async () => {
+  const response = await handler(eventNoBody, mockContext, mockCallback);
   expect(response).toBe(
     JSON.stringify({
       statusCode: 400,
       headers: { "Content-Type": "application/json" },
       body: {
-        message: "Bad request. Missing channel ID.",
+        message: "Bad request. Missing request body.",
       },
     })
   );
 });
 
 it("returns bad request for incorrect format of channel information", async () => {
-  const response = await handler(eventBadChannelId, mockContext, mockCallback);
+  const response = await handler(eventBadChannel, mockContext, mockCallback);
   expect(response).toBe(
     JSON.stringify({
       statusCode: 400,
       headers: { "Content-Type": "application/json" },
       body: {
-        message: "Bad request. Invalid channel ID.",
+        message: "Bad request. Invalid channel information.",
       },
     })
   );
