+++ src/test/addUser.test.ts
@@ -1,21 +1,22 @@
 import { APIGatewayProxyEventV2 } from "aws-lambda";
-import { handler } from "../lambdas/getUserSavedChannels";
+import { handler } from "@/lambdas/addUser";
 import { mockCallback, mockContext, mockEvent } from "./constants";
 
-const eventNoParams: APIGatewayProxyEventV2 = {
+const eventNoBody: APIGatewayProxyEventV2 = {
   ...mockEvent,
-  pathParameters: undefined,
+  body: undefined,
 };
 
-const eventBadSavedChannel: APIGatewayProxyEventV2 = {
+const eventBadUser: APIGatewayProxyEventV2 = {
   ...mockEvent,
-  pathParameters: {
-    channel_id: "1234",
-  },
+  body: JSON.stringify({
+    user_id: "123",
+    email: "test@gmail.com",
+  }),
 };
 
-it("returns bad request for missing parameters", async () => {
-  const response = await handler(eventNoParams, mockContext, mockCallback);
+it("returns bad request for missing body", async () => {
+  const response = await handler(eventNoBody, mockContext, mockCallback);
   expect(response).toBe(
     JSON.stringify({
       statusCode: 400,
@@ -28,11 +29,7 @@ it("returns bad request for missing parameters", async () => {
 });
 
 it("returns bad request for incorrect format of user information", async () => {
-  const response = await handler(
-    eventBadSavedChannel,
-    mockContext,
-    mockCallback
-  );
+  const response = await handler(eventBadUser, mockContext, mockCallback);
   expect(response).toBe(
     JSON.stringify({
       statusCode: 400,
