+++ src/test/deleteSavedChannel.test.ts
@@ -1,41 +1,45 @@
 import { APIGatewayProxyEventV2 } from "aws-lambda";
-import { handler } from "../lambdas/addChannel";
+import { handler } from "@/lambdas/deleteSavedChannel";
 import { mockCallback, mockContext, mockEvent } from "./constants";
 
-const eventNoBody: APIGatewayProxyEventV2 = {
+const eventNoParams: APIGatewayProxyEventV2 = {
   ...mockEvent,
-  body: undefined,
+  queryStringParameters: undefined,
 };
 
-const eventBadChannel: APIGatewayProxyEventV2 = {
+const eventBadSavedChannel: APIGatewayProxyEventV2 = {
   ...mockEvent,
-  body: JSON.stringify({
-    channelId: 123,
-    platform: "twitch",
-  }),
+  queryStringParameters: {
+    user_id: "1234",
+    channel: "1234",
+  },
 };
 
 it("returns bad request for missing body", async () => {
-  const response = await handler(eventNoBody, mockContext, mockCallback);
+  const response = await handler(eventNoParams, mockContext, mockCallback);
   expect(response).toBe(
     JSON.stringify({
       statusCode: 400,
       headers: { "Content-Type": "application/json" },
       body: {
-        message: "Bad request. Missing request body.",
+        message: "Bad request. Missing user and/or channel information.",
       },
     })
   );
 });
 
 it("returns bad request for incorrect format of channel information", async () => {
-  const response = await handler(eventBadChannel, mockContext, mockCallback);
+  const response = await handler(
+    eventBadSavedChannel,
+    mockContext,
+    mockCallback
+  );
   expect(response).toBe(
     JSON.stringify({
       statusCode: 400,
       headers: { "Content-Type": "application/json" },
       body: {
-        message: "Bad request. Invalid channel information.",
+        message: "Bad request. Invalid user and/or channel information.",
       },
     })
   );
