+++ src/test/deleteUser.test.ts
@@ -1,45 +1,40 @@
 import { APIGatewayProxyEventV2 } from "aws-lambda";
-import { handler } from "../lambdas/deleteSavedChannel";
+import { handler } from "@/lambdas/deleteUser";
 import { mockCallback, mockContext, mockEvent } from "./constants";
 
 const eventNoParams: APIGatewayProxyEventV2 = {
   ...mockEvent,
-  queryStringParameters: undefined,
+  pathParameters: undefined,
 };
 
-const eventBadSavedChannel: APIGatewayProxyEventV2 = {
+const eventBadUserId: APIGatewayProxyEventV2 = {
   ...mockEvent,
-  queryStringParameters: {
-    user_id: "1234",
-    channel: "1234",
+  pathParameters: {
+    channel_id: "1234",
   },
 };
 
-it("returns bad request for missing body", async () => {
+it("returns bad request for missing path params", async () => {
   const response = await handler(eventNoParams, mockContext, mockCallback);
   expect(response).toBe(
     JSON.stringify({
       statusCode: 400,
       headers: { "Content-Type": "application/json" },
       body: {
-        message: "Bad request. Missing user and/or channel information.",
+        message: "Bad request. Missing user ID.",
       },
     })
   );
 });
 
-it("returns bad request for incorrect format of channel information", async () => {
-  const response = await handler(
-    eventBadSavedChannel,
-    mockContext,
-    mockCallback
-  );
+it("returns bad request for incorrect format of user information", async () => {
+  const response = await handler(eventBadUserId, mockContext, mockCallback);
   expect(response).toBe(
     JSON.stringify({
       statusCode: 400,
       headers: { "Content-Type": "application/json" },
       body: {
-        message: "Bad request. Invalid user and/or channel information.",
+        message: "Bad request. Invalid user ID.",
       },
     })
   );
